more review comments have been addressed
